Now just need to get S3 working

+++ thirdai_python_package/telemetry/telemetry_start_and_stop.py
@@ -2,9 +2,9 @@ import atexit
 import pathlib
 import subprocess
 import sys
+import uuid
 from typing import Optional
 
-import pandas as pd
 import thirdai
 
 daemon_path = pathlib.Path(__file__).parent.resolve() / "telemetry_daemon.py"
@@ -13,10 +13,17 @@ background_process = None
 
 BACKGROUND_THREAD_TIMEOUT_SECONDS = 0.5
 
+UUID = uuid.uuid4().hex
+
 
 def kill_background_process():
     global background_process
     if background_process != None:
+        poll = background_process.poll()
+        if poll is None:
+            raise ValueError(
+                f"Telemetry process terminated early with exit code {poll}"
+            )
         background_process.terminate()
         try:
             background_process.wait(timeout=BACKGROUND_THREAD_TIMEOUT_SECONDS)
@@ -32,7 +39,7 @@ wrapped_start_method = thirdai._thirdai.telemetry.start
 wrapped_stop_method = thirdai._thirdai.telemetry.stop
 
 
-def start(port: Optional[int] = None, write_location: Optional[str] = None):
+def start(port: Optional[int] = None, write_dir: Optional[str] = None):
     global background_process
     if background_process != None:
         raise RuntimeError(
@@ -40,23 +47,28 @@ def start(port: Optional[int] = None, write_location: Optional[str] = None):
         )
 
     if port:
-        url = wrapped_start_method(port)
+        telemetry_url = wrapped_start_method(port)
     else:
-        url = wrapped_start_method()
+        telemetry_url = wrapped_start_method()
 
-    if write_location != None:
-        # Could also try using os.fork
-        python_executable = sys.executable
-        background_process = subprocess.Popen(
-            [
-                python_executable,
-                str(daemon_path.resolve()),
-                "--telemetry_url",
-                "http://" + url + "/metrics",
-                "--push_location",
-                write_location,
-            ]
-        )
+    if write_dir == None:
+        return telemetry_url
+
+    # Could also try using os.fork
+    python_executable = sys.executable
+    push_location = str(pathlib.Path(write_dir) / ("telemetry-" + UUID))
+    background_process = subprocess.Popen(
+        [
+            python_executable,
+            str(daemon_path.resolve()),
+            "--telemetry_url",
+            telemetry_url,
+            "--push_location",
+            push_location,
+        ]
+    )
+
+    return push_location
 
 
 def stop():
